backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
sqlmodel==0.0.22
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
websockets==13.1
httpx==0.28.1
pydantic-settings==2.7.1
